Lighting + movement improvements
I added the limited vision and fixed some issues with the player and monster movement. I also made it so that u can move with either wasd or the arrow keys as well as the ability to sprint. The items aren't ready yet so ill work on those next. Let me know if u want me to cahnge anything