Temporal Procedural Stage Generation
Added stage generation system that can be used for prototype.

There is no door yet and corridor might be sort of off, so its still temporary.

Once the yellow line connecting all the red squares appear on screen, press tab key to generate interactive stage.

Make sure the player is within the red square when generating stage, as outside of map is filled with collision.